Make changes to User repo and expenses.
Still need to complete expenses.

+++ server/src/controllers/expenseController.ts
@@ -4,40 +4,42 @@ import { ExpenseRepository } from "../repositories/ExpenseRepository";
 export const ExpenseController = {
     async getAllExpenses(req: Request, res: Response) {
         const result = await ExpenseRepository.getAllExpenses();
-        res.status(200).json(result);
+        if (result) {
+            res.status(200).json(result.rows);
+        }
     },
 
     async getAllUserExpenses(req: Request, res: Response) {
-        const id: Number = Number(req.params.id);
+        const id: string = req.params.id;
         
         const result = await ExpenseRepository.getAllUserExpenses(id);
-        if (result) {
-            res.status(200).json({success: true, expense: result});
+        if (result?.rowLength != null && result.rowLength > 0) {
+            res.status(200).json({success: true, expense: result.rows});
         } else {
-            res.status(404).json({success: false, message:"User and/or Expenses not found", expense: result});
+            res.status(404).json({success: false, message:"User and/or Expenses not found", expense: null});
         }
     },
 
     async getAllUserExpensesByCategory(req: Request, res: Response) {
-        const userid: Number = Number(req.params.id);
-        const categoryid: Number = Number(req.params.categoryid);
+        const userid: string = req.params.id;
+        const categoryid: string = req.params.categoryid;
         
         const result = await ExpenseRepository.getAllUserExpensesByCategory(userid, categoryid);
-        if (result) {
-            res.status(200).json({success: true, expense: result});
+        if (result?.rowLength != null && result.rowLength > 0) {
+            res.status(200).json({success: true, expense: result.rows});
         } else {
-            res.status(404).json({success: false, message:"User and/or Category not found", expense: result});
+            res.status(404).json({success: false, message:"User and/or Category not found", expense: null});
         }
     },
 
     async getExpenseById(req: Request, res: Response) {
-        const id: Number = Number(req.params.id);
+        const id: string = req.params.id;
         
         const result = await ExpenseRepository.getExpenseById(id);
-        if (result) {
-            res.status(200).json({success: true, expense: result});
+        if (result?.rowLength != null && result.rowLength > 0) {
+            res.status(200).json({success: true, expense: result.rows});
         } else {
-            res.status(404).json({success: false, message:"Expense not found", expense: result});
+            res.status(404).json({success: false, message:"Expense not found", expense: null});
         }
     },
 
@@ -51,69 +53,81 @@ export const ExpenseController = {
     },
 
     async updateExpense(req: Request, res: Response) {
-        const id: Number = Number(req.params.id);
+        const id: string = req.params.id;
         const updatedExpense = req.body;
-        const result = await ExpenseRepository.updateExpense(id, updatedExpense);
-        if (result) {
-            res.status(204).json({success: true});
+
+        const checkExist = await ExpenseRepository.getExpenseById(id);
+        if (checkExist?.rowLength != null && checkExist.rowLength > 0) {
+            const result = await ExpenseRepository.updateExpense(id, updatedExpense);
+            if (result) {
+                res.status(200).json({success: true});
+            } else {
+                res.status(400).json({success: false, message:"Error"});
+            }
         } else {
             res.status(404).json({success: false, message:"Expense not found"});
         }
     },
 
     async deleteExpense(req: Request, res: Response) {
-        const id: Number = Number(req.params.id);
-        const result = await ExpenseRepository.deleteExpense(id);
-        if (result) {
-            res.status(200).json({success: true});
+        const id: string = req.params.id;
+        
+        const checkExist = await ExpenseRepository.getExpenseById(id);
+        if (checkExist?.rowLength != null && checkExist.rowLength > 0) {
+            const result = await ExpenseRepository.deleteExpense(id);
+            if (result) {
+                res.status(200).json({success: true});
+            } else {
+                res.status(400).json({success: false, message:"Error", category: null});
+            }
         } else {
             res.status(404).json({success: false, message:"Expense not found"});
         }
     },
     
     async getAvgAmount(req: Request, res: Response) {
-        const id: Number = Number(req.params.id);
+        const id: string = req.params.id;
 
         const result = await ExpenseRepository.getAvgAmount(id);
-        if (result) {
-            res.status(200).json({success: true, expense: result});
+        if (result?.rowLength != null && result.rowLength > 0) {
+            res.status(200).json({success: true, expense: result.rows});
         } else {
-            res.status(404).json({success: false, message:"User and/or Expenses not found", expense: result});
+            res.status(404).json({success: false, message:"User and/or Expenses not found", expense: null});
         }
     },
 
     async getSumAmount(req: Request, res: Response) {
-        const id: Number = Number(req.params.id);
+        const id: string = req.params.id;
         
         const result = await ExpenseRepository.getSumAmount(id);
-        if (result) {
-            res.status(200).json({success: true, expense: result});
+        if (result?.rowLength != null && result.rowLength > 0) {
+            res.status(200).json({success: true, expense: result.rows});
         } else {
-            res.status(404).json({success: false, message:"User and/or Expenses not found", expense: result});
+            res.status(404).json({success: false, message:"User and/or Expenses not found", expense: null});
         }
     },
 
     async getAvgAmountByCategory(req: Request, res: Response) {
-        const userid: Number = Number(req.params.id);
-        const categoryid: Number = Number(req.params.categoryid);
+        const userid: string = req.params.id;
+        const categoryid: string = req.params.categoryid;
         
         const result = await ExpenseRepository.getAvgAmountByCategory(userid, categoryid);
-        if (result) {
-            res.status(200).json({success: true, expense: result});
+        if (result?.rowLength != null && result.rowLength > 0) {
+            res.status(200).json({success: true, expense: result.rows});
         } else {
-            res.status(404).json({success: false, message:"User and/or Category not found", expense: result});
+            res.status(404).json({success: false, message:"User and/or Category not found", expense: null});
         }
     },
     
     async getSumAmountByCategory(req: Request, res: Response) {
-        const userid: Number = Number(req.params.id);
-        const categoryid: Number = Number(req.params.categoryid);
+        const userid: string = req.params.id;
+        const categoryid: string = req.params.categoryid;
         
         const result = await ExpenseRepository.getSumAmountByCategory(userid, categoryid);
-        if (result) {
-            res.status(200).json({success: true, expense: result});
+        if (result?.rowLength != null && result.rowLength > 0) {
+            res.status(200).json({success: true, expense: result.rows});
         } else {
-            res.status(404).json({success: false, message:"User and/or Category not found", expense: result});
+            res.status(404).json({success: false, message:"User and/or Category not found", expense: null});
         }
     },
 };
